refactor(react-langgraph): adopt createRuntimeExtras and split the runtime file
replace the hand-rolled Symbol-extras brand + asLangGraphRuntimeExtras guard with
the shared createRuntimeExtras helper from @assistant-ui/core/internal; move the
five accessor hooks into hooks.ts, UseLangGraphRuntimeOptions + the extras type
into types.ts, the outbound getMessageContent mapper into convertLangChainMessages.ts,
and the pure getPendingToolCalls/truncate/filter helpers into messageHelpers.ts.
the accumulator, generative-UI reconcile effect, queue, and load wiring are
unchanged, and the public barrel is preserved exactly (no export removed). also
adds a public-surface guardrail to AGENTS.md.
