+++ .changeset/langgraph-runtime-extras.md
@@ -1,0 +1,5 @@
+---
+"@assistant-ui/react-langgraph": patch
+---
+
+refactor: adopt the shared createRuntimeExtras helper and split useLangGraphRuntime.ts into hooks/types/converter/helper modules, with no public API or behavior change
