+++ AGENTS.md
@@ -52,6 +52,7 @@ A framework adapter (`@assistant-ui/react-*`) maps a provider onto a core runtim
 - **Core runtime.** Build on `useExternalStoreRuntime` (messages derived from an external source) or `useLocalRuntime` plus a `ChatModelAdapter` (no provider-side thread state, like `react-data-stream`), and wrap it in `useRemoteThreadListRuntime` for multi-thread support.
 - **State exposure.** Expose runtime state to accessor hooks with `createRuntimeExtras` from `@assistant-ui/core/internal`, not a hand-rolled `Symbol` brand and guard.
 - **Standard files.** `use<Name>Runtime.ts` (orchestration only), `hooks.ts` (accessor and action hooks), a pure `convertMessages.ts` (both directions), and `types.ts`; add a `<Name>ThreadController.ts` plus a pure `reduce<Name>ThreadState` reducer when the adapter owns thread state, and a `./server` or `./node` subpath entry when the protocol owns the wire.
+- **Public surface.** A new adapter's `index.ts` should re-export only the entry hook, accessor hooks, converters, and public types. For an already-published package, treat the barrel as append-only: re-point moved exports to their new files, but never remove an export that has shipped (a breaking change for npm consumers the in-repo audit cannot see), even an unused-looking type.
 - **Tests.** Colocate them beside each module, covering the converter both ways, the reducer or controller, and each accessor hook.
 - **Do not introduce.** A state holder named `*ThreadRuntimeCore`, a `notifyUpdate` plus version-counter re-render hack (bridge non-React state with `useSyncExternalStore`), `Object.create` method grafting, or monkeypatching the caller's objects. The legacy `react-a2a` and `react-ag-ui` predate this and are being migrated.
