+++ packages/react-langgraph/src/runtimeExtras.ts
@@ -1,0 +1,6 @@
+import { createRuntimeExtras } from "@assistant-ui/core/internal";
+import type { LangGraphRuntimeExtras } from "./types";
+
+export const langGraphExtras = createRuntimeExtras<LangGraphRuntimeExtras>(
+  "useLangGraphRuntime",
+);
