+++ packages/react-langgraph/src/useLangGraphRuntime.ts
@@ -5,35 +5,13 @@ import {
   useState,
   useSyncExternalStore,
 } from "react";
-import type {
-  LangChainMessage,
-  LangChainToolCall,
-  LangGraphTupleMetadata,
-  OnMessageChunkCallback,
-  OnValuesEventCallback,
-  OnUpdatesEventCallback,
-  OnSubgraphValuesEventCallback,
-  OnSubgraphUpdatesEventCallback,
-  OnCustomEventCallback,
-  OnErrorEventCallback,
-  OnSubgraphErrorEventCallback,
-  OnInfoEventCallback,
-  OnMetadataEventCallback,
-  UIMessage,
-} from "./types";
+import type { LangChainMessage, UIMessage } from "./types";
+import type { UseLangGraphRuntimeOptions } from "./types";
 import {
-  getExternalStoreMessages,
   pickExternalStoreSharedOptions,
   createMessageQueue,
   type MessageQueueController,
-  type ThreadMessage,
-  type AttachmentAdapter,
   type AppendMessage,
-  type DictationAdapter,
-  type ExternalStoreSharedOptions,
-  type FeedbackAdapter,
-  type RealtimeVoiceAdapter,
-  type SpeechSynthesisAdapter,
 } from "@assistant-ui/core";
 import type { ToolExecutionStatus } from "@assistant-ui/core";
 import type { QueueItemState } from "@assistant-ui/core/store";
@@ -44,320 +22,27 @@ import {
   useExternalMessageConverter,
   useExternalStoreRuntime,
 } from "@assistant-ui/core/react";
-import { useAui, useAuiState } from "@assistant-ui/store";
-import type { AssistantCloud } from "assistant-cloud";
-import type { RemoteThreadListAdapter } from "@assistant-ui/core";
-import { convertLangChainMessages } from "./convertLangChainMessages";
+import { useAui } from "@assistant-ui/store";
 import {
-  type LangGraphCommand,
-  type LangGraphInterruptState,
+  convertLangChainMessages,
+  getMessageContent,
+} from "./convertLangChainMessages";
+import {
   type LangGraphSendMessageConfig,
-  type LangGraphStreamCallback,
   useLangGraphMessages,
 } from "./useLangGraphMessages";
 import { appendLangChainChunk } from "./appendLangChainChunk";
 import { useLangGraphStreamingTiming } from "./useLangGraphStreamingTiming";
 import { bufferToolResult } from "./bufferToolResults";
-
-const getPendingToolCalls = (messages: LangChainMessage[]) => {
-  const pendingToolCalls = new Map<string, LangChainToolCall>();
-  for (const message of messages) {
-    if (message.type === "ai") {
-      for (const toolCall of message.tool_calls ?? []) {
-        pendingToolCalls.set(toolCall.id, toolCall);
-      }
-    }
-    if (message.type === "tool") {
-      pendingToolCalls.delete(message.tool_call_id);
-    }
-  }
-
-  return [...pendingToolCalls.values()];
-};
-
-const getMessageContent = (msg: AppendMessage) => {
-  const allContent = [
-    ...msg.content,
-    ...(msg.attachments?.flatMap((a) => a.content) ?? []),
-  ];
-
-  const hasNonText = allContent.some(
-    (part) => part.type === "file" || part.type === "image",
-  );
-  const hasText = allContent.some((part) => part.type === "text");
-  if (hasNonText && !hasText) {
-    allContent.unshift({ type: "text", text: " " });
-  }
-
-  const content = allContent.map((part) => {
-    const type = part.type;
-    switch (type) {
-      case "text":
-        return { type: "text" as const, text: part.text };
-      case "image":
-        return { type: "image_url" as const, image_url: { url: part.image } };
-      case "file":
-        return {
-          type: "file" as const,
-          data: part.data,
-          mime_type: part.mimeType,
-          metadata: {
-            filename: part.filename ?? "file",
-          },
-          source_type: "base64" as const,
-        };
-
-      case "tool-call":
-        throw new Error("Tool call appends are not supported.");
-
-      default: {
-        const _exhaustiveCheck:
-          | "reasoning"
-          | "source"
-          | "audio"
-          | "data"
-          | "generative-ui" = type;
-        throw new Error(
-          `Unsupported append message part type: ${_exhaustiveCheck}`,
-        );
-      }
-    }
-  });
-
-  if (content.length === 1 && content[0]?.type === "text") {
-    return content[0].text ?? "";
-  }
-
-  return content;
-};
-
-const symbolLangGraphRuntimeExtras = Symbol("langgraph-runtime-extras");
-type LangGraphRuntimeExtras = {
-  [symbolLangGraphRuntimeExtras]: true;
-  send: (
-    messages: LangChainMessage[],
-    config: LangGraphSendMessageConfig,
-  ) => Promise<void>;
-  interrupt: LangGraphInterruptState | undefined;
-  messageMetadata: Map<string, LangGraphTupleMetadata>;
-  uiMessages: readonly UIMessage[];
-};
-
-const asLangGraphRuntimeExtras = (extras: unknown): LangGraphRuntimeExtras => {
-  if (
-    typeof extras !== "object" ||
-    extras == null ||
-    !(symbolLangGraphRuntimeExtras in extras)
-  )
-    throw new Error(
-      "This method can only be called when you are using useLangGraphRuntime",
-    );
-
-  return extras as LangGraphRuntimeExtras;
-};
-
-export const useLangGraphInterruptState = () => {
-  const interrupt = useAuiState((s) => {
-    const extras = s.thread.extras;
-    if (!extras) return undefined;
-    return asLangGraphRuntimeExtras(extras).interrupt;
-  });
-  return interrupt;
-};
-
-export const useLangGraphSend = () => {
-  const aui = useAui();
-
-  return (messages: LangChainMessage[], config: LangGraphSendMessageConfig) => {
-    const extras = aui.thread().getState().extras;
-    const { send } = asLangGraphRuntimeExtras(extras);
-    return send(messages, config);
-  };
-};
-
-export const useLangGraphSendCommand = () => {
-  const send = useLangGraphSend();
-  return (command: LangGraphCommand) => send([], { command });
-};
-
-export const useLangGraphMessageMetadata = () => {
-  const messageMetadata = useAuiState((s) => {
-    const extras = s.thread.extras;
-    if (!extras) return new Map<string, LangGraphTupleMetadata>();
-    return asLangGraphRuntimeExtras(extras).messageMetadata;
-  });
-  return messageMetadata;
-};
-
-const EMPTY_UI_MESSAGES: readonly UIMessage[] = Object.freeze([]);
+import { langGraphExtras } from "./runtimeExtras";
+import {
+  filterUIMessagesBySurvivingIds,
+  getPendingToolCalls,
+  truncateLangChainMessages,
+} from "./messageHelpers";
 
 const EMPTY_QUEUE_ITEMS: readonly QueueItemState[] = Object.freeze([]);
 const subscribeNoop = () => () => {};
-
-export const useLangGraphUIMessages = () => {
-  return useAuiState((s) => {
-    const extras = s.thread.extras;
-    if (!extras) return EMPTY_UI_MESSAGES;
-    return asLangGraphRuntimeExtras(extras).uiMessages;
-  });
-};
-
-export type UseLangGraphRuntimeOptions = ExternalStoreSharedOptions & {
-  autoCancelPendingToolCalls?: boolean | undefined;
-  /**
-   * When true, renders the Cancel button in the composer and aborts the
-   * `AbortController` whose signal is exposed to your `stream` callback
-   * as `config.abortSignal`.
-   */
-  unstable_allowCancellation?: boolean | undefined;
-  /**
-   * Opt in to message queuing: a message sent during a run is held in
-   * `composer.queue` and sent once the run settles. Steering runs it next.
-   */
-  unstable_enableMessageQueue?: boolean | undefined;
-  stream: LangGraphStreamCallback<LangChainMessage>;
-  /**
-   * State key under which LangGraph's `typed_ui` writes Generative UI
-   * messages in the graph state. Must match the `stateKey` option passed to
-   * `typedUi(config, { stateKey })` on the server. Defaults to `"ui"`.
-   */
-  uiStateKey?: string;
-  /**
-   * Resolves a checkpoint ID for a given thread and message history.
-   * When provided, enables message editing (onEdit) and regeneration (onReload).
-   * The checkpoint ID is passed to the stream callback for server-side forking.
-   */
-  getCheckpointId?: (
-    threadId: string,
-    parentMessages: LangChainMessage[],
-  ) => Promise<string | null>;
-  load?: (
-    threadId: string,
-    config?: { signal: AbortSignal },
-  ) => Promise<{
-    messages: LangChainMessage[];
-    interrupts?: LangGraphInterruptState[];
-    /**
-     * Persisted LangSmith Generative UI messages for this thread, typically
-     * read from `state.values[uiStateKey]` returned by the LangGraph SDK's
-     * `client.threads.getState()`. Defaults to an empty list.
-     */
-    uiMessages?: UIMessage[];
-  }>;
-  create?: () => Promise<{
-    externalId: string;
-  }>;
-  delete?: (threadId: string) => Promise<void>;
-  adapters?:
-    | {
-        attachments?: AttachmentAdapter;
-        speech?: SpeechSynthesisAdapter;
-        dictation?: DictationAdapter;
-        voice?: RealtimeVoiceAdapter;
-        feedback?: FeedbackAdapter;
-      }
-    | undefined;
-  eventHandlers?:
-    | {
-        /**
-         * Called for each message chunk received from messages-tuple streaming,
-         * with the chunk and its associated metadata
-         */
-        onMessageChunk?: OnMessageChunkCallback;
-        /**
-         * Called when top-level values events are received from the LangGraph stream.
-         * Subgraph values are routed to `onSubgraphValues`.
-         */
-        onValues?: OnValuesEventCallback;
-        /**
-         * Called when top-level updates events are received from the LangGraph stream.
-         * Subgraph updates are routed to `onSubgraphUpdates`.
-         */
-        onUpdates?: OnUpdatesEventCallback;
-        /** Called when a subgraph (namespaced) values event is received. */
-        onSubgraphValues?: OnSubgraphValuesEventCallback;
-        /** Called when a subgraph (namespaced) updates event is received. */
-        onSubgraphUpdates?: OnSubgraphUpdatesEventCallback;
-        /**
-         * Called when metadata is received from the LangGraph stream
-         */
-        onMetadata?: OnMetadataEventCallback;
-        /**
-         * Called when informational messages are received from the LangGraph stream
-         */
-        onInfo?: OnInfoEventCallback;
-        /**
-         * Called when errors occur during LangGraph stream processing.
-         * Fires for both top-level and subgraph errors; subgraph errors
-         * additionally trigger `onSubgraphError` with the namespace.
-         */
-        onError?: OnErrorEventCallback;
-        /** Called when a subgraph (namespaced) error event is received, in addition to `onError`. */
-        onSubgraphError?: OnSubgraphErrorEventCallback;
-        /**
-         * Called when custom events are received from the LangGraph stream
-         */
-        onCustomEvent?: OnCustomEventCallback;
-      }
-    | undefined;
-  /**
-   * Register data renderers for Generative UI components.
-   *
-   * `renderers` maps a `ui_message` name to a static component.
-   * `fallback` handles any name without a static match — use this for
-   * dynamic loading (e.g. LangSmith's `LoadExternalComponent`).
-   */
-  uiComponents?:
-    | {
-        fallback?: DataMessagePartComponent;
-        renderers?: Record<string, DataMessagePartComponent>;
-      }
-    | undefined;
-  cloud?: AssistantCloud | undefined;
-  /**
-   * A `RemoteThreadListAdapter` to use instead of the cloud adapter. Provide
-   * this to back the thread list with a custom store (e.g. LangGraph
-   * `client.threads.search()`) so pre-existing LangGraph thread ids appear in
-   * the UI and can be switched between without assistant-cloud.
-   *
-   * When provided, `cloud`, `create`, and `delete` are ignored — the adapter
-   * owns the full thread list lifecycle. The `externalId` returned by the
-   * adapter's `list()` / `initialize()` is what the `load` callback receives.
-   */
-  unstable_threadListAdapter?: RemoteThreadListAdapter | undefined;
-};
-
-const truncateLangChainMessages = (
-  threadMessages: readonly ThreadMessage[],
-  parentId: string | null,
-): LangChainMessage[] => {
-  if (parentId === null) return [];
-  const parentIndex = threadMessages.findIndex((m) => m.id === parentId);
-  if (parentIndex === -1) return [];
-  const truncated: LangChainMessage[] = [];
-  for (let i = 0; i <= parentIndex && i < threadMessages.length; i++) {
-    truncated.push(
-      ...getExternalStoreMessages<LangChainMessage>(threadMessages[i]!),
-    );
-  }
-  return truncated;
-};
-
-const filterUIMessagesBySurvivingIds = (
-  uiMessages: readonly UIMessage[],
-  survivingMessages: readonly LangChainMessage[],
-): UIMessage[] => {
-  const survivingIds = new Set<string>();
-  for (const m of survivingMessages) {
-    if (m.id) survivingIds.add(m.id);
-  }
-  return uiMessages.filter((ui) => {
-    const parentId = ui.metadata?.message_id;
-    // orphans (no message_id) represent global UI, cleared only via delete_ui_message
-    if (!parentId) return true;
-    return survivingIds.has(parentId);
-  });
-};
 
 const useLangGraphRuntimeImpl = (options: UseLangGraphRuntimeOptions) => {
   const {
@@ -590,13 +275,12 @@ const useLangGraphRuntimeImpl = (options: UseLangGraphRuntimeOptions) => {
       speech,
       voice,
     },
-    extras: {
-      [symbolLangGraphRuntimeExtras]: true,
+    extras: langGraphExtras.provide({
       interrupt,
       messageMetadata,
       uiMessages,
       send: handleSendMessage,
-    } satisfies LangGraphRuntimeExtras,
+    }),
     onNew: runUserMessage,
     ...(queueController && { queue: queueController.adapter }),
     onAddToolResult: async ({
